requirements.txt
# === Основное ===
streamlit==1.39.0          # Интерфейс
opencv-python==4.10.0.84   # Работа с изображениями
numpy==1.26.4              # Обработка массивов
pillow==10.4.0             # Работа с изображениями PIL
matplotlib==3.9.2          # Графики для аналитики
pandas==2.2.3              # Таблицы и отчёты

# === Torch и YOLO ===
torch>=2.3.0
torchvision>=0.18.0
ultralytics==8.3.30        # YOLOv8/v9 — для работы с моделью

# === Отчёты и PDF ===
reportlab==4.2.5           # Генерация PDF отчётов

# === Утилиты ===
tqdm==4.66.5               # Прогресс-бары
scikit-image==0.24.0       # Для возможных фильтров / XAI
